+++ app/services/whatsapp/incoming_message_base_service.rb
@@ -48,10 +48,12 @@ class Whatsapp::IncomingMessageBaseService
 
   def process_statuses
     status = @processed_params[:statuses].first
-    return unless find_message_by_source_id(status[:id])
-
-    update_whatsapp_identifiers_from_status(status)
-    update_message_with_status(@message, status)
+    if find_message_by_source_id(status[:id])
+      update_whatsapp_identifiers_from_status(status)
+      update_message_with_status(@message, status)
+    else
+      CampaignDashboard::ForwardWhatsappStatusService.new(inbox: inbox, status: status).perform
+    end
   rescue ArgumentError => e
     Rails.logger.error "Error while processing whatsapp status update #{e.message}"
   end